requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMuPDF==1.25.5
pytesseract==0.3.13
pywin32-ctypes==0.2.3
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
